+++ layouts/MainLayout.js
@@ -1,16 +1,19 @@
 import Header from '../components/Header';
 import Footer from '../components/Footer';
+import { AuthProvider } from '../utils/auth/authContext';
 
 // set layout for every page, eliminating the need to import common components individually
 const MainLayout = (Page) => {
   function Wrapper(props) {
     return (
       <div>
-        <Header />
-        <main className='main-content'>
-          <Page {...props} />
-        </main>
-        <Footer />
+        <AuthProvider>
+          <Header />
+          <main className='main-content'>
+            <Page {...props} />
+          </main>
+          <Footer />
+        </AuthProvider>
       </div>
     );
   }